synapse_backend/requirements.txt
fastapi[all]
uvicorn[standard]
sqlalchemy
python-dotenv
google-generativeai
langchain
langgraph
langchain-google-genai
chromadb
pypdf
python-docx
langchain-core
passlib[bcrypt]
python-jose[cryptography]
google-cloud-speech
google-cloud-texttospeech